gnome-session: Bind more LC_* variables to REGION
This adds both LC_ADDRESS and LC_TELEPHONE to the list of variables that
are set from the configured user region.

Also change the list to be in the same order as printed by the "locale"
command and add comments for entries that we are not setting.

Origin: commit 6e76dc3761caac4b1b211c33538d86457aaa5687